fix error

+++ userbot/plugins/الفحص.py
@@ -43,10 +43,10 @@ async def amireallyalive(event):
 
 
 @jmthon.ar_cmd(
-    pattern="فحص",
+    pattern="فحص$",
     command=("فحص", plugin_category),
     info={
-        "header": "To check bot's alive status via inline mode",
+        "header": "To check bot's alive status",
         "options": "To show media in this cmd you need to set ALIVE_PIC with media link, get this by replying the media by .tgm",
         "usage": [
             "{tr}alive",
@@ -54,18 +54,39 @@ async def amireallyalive(event):
     },
 )
 async def amireallyalive(event):
-    "A kind of showing bot details by your inline bot"
+    "A kind of showing bot details"
     reply_to_id = await reply_id(event)
+    uptime = await get_readable_time((time.time() - StartTime))
+    _, check_sgnirts = check_data_base_heal_th()
     EMOJI = gvarstatus("ALIVE_EMOJI") or "  - "
-    cat_caption = f"**بـوت جـمثون يعـمل بـنجـاح**\n"
-    cat_caption += f"**{EMOJI} نسخۿ التيليثون :** `{version.__version__}\n`"
-    cat_caption += f"**{EMOJI} نسخۿ جمثون :** `{catversion}`\n"
-    cat_caption += f"**{EMOJI} نسخۿ البـايثون :** `{python_version()}\n`"
-    cat_caption += f"**{EMOJI} الـمالك:** {mention}\n"
-    results = await event.client.inline_query(Config.TG_BOT_USERNAME, cat_caption)
-    await results[0].click(event.chat_id, reply_to=reply_to_id, hide_via=True)
-    await event.delete()
-
+    CUSTOM_ALIVE_TEXT = gvarstatus("ALIVE_TEXT") or "⇝ WELCOME TO JMTHON 𝄵 ⇜"
+    CAT_IMG = gvarstatus("ALIVE_PIC")
+    if CAT_IMG:
+        CAT = [x for x in CAT_IMG.split()]
+        A_IMG = list(CAT)
+        PIC = random.choice(A_IMG)
+        cat_caption = f"**{CUSTOM_ALIVE_TEXT}**\n\n"
+        cat_caption += f"**{EMOJI} قاعدۿ البيانات :** `تـعمل بنـجاح`\n"
+        cat_caption += f"**{EMOJI} نسخۿ تليثون :** `{version.__version__}\n`"
+        cat_caption += f"**{EMOJI} نسخـۿ جـمثون :** `{catversion}`\n"
+        cat_caption += f"**{EMOJI} نسخـۿ البايثون :** `{python_version()}\n`"
+        cat_caption += f"**{EMOJI} الوقت :** `{uptime}\n`"
+        cat_caption += f"**{EMOJI} المنشئ:** {mention}\n"
+        await event.client.send_file(
+            event.chat_id, PIC, caption=cat_caption, reply_to=reply_to_id
+        )
+        await event.delete()
+    else:
+        await edit_or_reply(
+            event,
+            f"**{CUSTOM_ALIVE_TEXT}**\n\n"
+            f"**{EMOJI} قاعدۿ البيانات :** `تـعمل بنجـاح`\n"
+            f"**{EMOJI} نسخۿ تليثون :** `{version.__version__}\n`"
+            f"**{EMOJI} نسخـۿ جـمثون :** `{catversion}`\n"
+            f"**{EMOJI} نسخـۿ البايثون :** `{python_version()}\n`"
+            f"**{EMOJI} الوقت :** `{uptime}\n`"
+            f"**{EMOJI} المنشئ:** {mention}\n",
+        )
 
 @jmthon.tgbot.on(CallbackQuery(data=re.compile(b"stats")))
 async def on_plug_in_callback_query_handler(event):
